Types cleanup
