--- packages/echo-base/src/types/eventHub.ts
@@ -1,9 +1,0 @@
-export interface EchoEventHub {
-    emit<T>(key: string, payload: T): void;
-    subscribe<T>(key: string, handler: (payload: T) => void): () => void;
-    subscribeMany<T>(keys: Array<string | EchoEvents>, handler: (payload: T) => void): () => void;
-}
-
-export enum EchoEvents {
-    PlantChanged = 'plantChanged'
-}
